Refactored magazines to support MVVM
- Created a ViewModel class for MagazinesList
- Moved all logic into the ViewModel class
